🔀 Squashed commit of the following:
commit 180c0ff689d68f4ab50eafdf7cf1b8dfde3a4cb6
Author: vgierlowski <victor.gierlowski@etu.unistra.fr>
Date:   Wed Feb 25 21:44:03 2026 +0100

    🔀  Merge Tower into main
    Squashed commit of the following:

    commit 292a540bb4504dfc62fedac59abdec18e1978141
    Author: vgierlowski <victor.gierlowski@etu.unistra.fr>
    Date:   Tue Feb 24 23:37:33 2026 +0100

        ✨ Refactor Enemy and Projectile classes to use float types; implement Target system for projectiles

    commit 692d69b9ca0e371d36aedcb2521c366336ccbc82
    Author: vgierlowski <victor.gierlowski@etu.unistra.fr>
    Date:   Tue Feb 24 21:28:57 2026 +0100

        ✨ Enhance Entity, Projectile, and Tower classes with new constructors and drawing functionality; refactor sprite creation methods

    commit 3d0da137a1afe5fe36aef43a349ea019664a1465
    Author: vgierlowski <victor.gierlowski@etu.unistra.fr>
    Date:   Tue Feb 24 21:03:09 2026 +0100

        ✨ Implement drawing functionality for Entity and enhance Projectile and Tower classes with new constructors and augment handling

    commit d22bd76e1af7fc04826a1fd2e04297ade97c1d1d
    Author: vgierlowski <victor.gierlowski@etu.unistra.fr>
    Date:   Fri Feb 20 11:32:09 2026 +0100

        🔧 Fix warning and swapped to float

    commit 62fcac305123cac473ad9ad0687219c2baecc014
    Author: vgierlowski <victor.gierlowski@etu.unistra.fr>
    Date:   Fri Feb 20 09:49:13 2026 +0100

        🚧 base augment projectile tower, postfix prefix

commit c972e6f38c521220cc8dd653eebf9eddcc2b47d0
Merge: f7c76da d67c09a
Author: vgierlowski <victor.gierlowski@etu.unistra.fr>
Date:   Wed Feb 25 21:37:01 2026 +0100

    🔀 Merge branch core

commit f7c76da56d59f6aeb127b85cf6d1f8059c4264ed
Merge: 8395271 fef8e96
Author: vgierlowski <victor.gierlowski@etu.unistra.fr>
Date:   Wed Feb 18 20:05:45 2026 +0100

    Merge branch 'DrawingEntity'

+++ src/Sprites/Sprite.cpp
@@ -23,7 +23,7 @@ Sprite::~Sprite(){}
 void Sprites::Sprite::draw(SDL_Renderer *win, float deltaTime, Point offset, float scale, float rot) {
     SDL_SetRenderDrawColor(win,100,255,55,255);
     offset += position_;
-    SDL_Rect objet = {offset.getX(),offset.getY(),21,21};
+    SDL_Rect objet = {(int)offset.getX(),(int)offset.getY(),21,21};
     SDL_RenderFillRect(win,&objet);
 }
